Updated speed to double
git-svn-id: https://svn.imm.dtu.dk/se2/svn/e13-groupA@554 2c9cc009-7d77-48f4-a239-25ca0ae51001

+++ src/dk.dtu.se2.animation/src/dk/dtu/se2/animation/Move.java
@@ -29,12 +29,12 @@ public interface Move extends Animation {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Speed</em>' attribute.
-	 * @see #setSpeed(int)
+	 * @see #setSpeed(double)
 	 * @see dk.dtu.se2.animation.AnimationPackage#getMove_Speed()
 	 * @model
 	 * @generated
 	 */
-	int getSpeed();
+	double getSpeed();
 
 	/**
 	 * Sets the value of the '{@link dk.dtu.se2.animation.Move#getSpeed <em>Speed</em>}' attribute.
@@ -44,6 +44,6 @@ public interface Move extends Animation {
 	 * @see #getSpeed()
 	 * @generated
 	 */
-	void setSpeed(int value);
+	void setSpeed(double value);
 
 } // Move
